+++ common/src/main/java/org/orecruncher/dsurround/Client.java
@@ -9,7 +9,6 @@ import net.minecraft.server.packs.PackType;
 import org.orecruncher.dsurround.commands.Commands;
 import org.orecruncher.dsurround.config.libraries.*;
 import org.orecruncher.dsurround.config.libraries.impl.*;
-import org.orecruncher.dsurround.effects.particles.ParticleSheets;
 import org.orecruncher.dsurround.gui.overlay.OverlayManager;
 import org.orecruncher.dsurround.gui.keyboard.KeyBindings;
 import org.orecruncher.dsurround.lib.GameUtils;
@@ -24,6 +23,8 @@ import org.orecruncher.dsurround.lib.logging.ModLog;
 import org.orecruncher.dsurround.eventing.ClientState;
 import org.orecruncher.dsurround.lib.registry.ReloadListener;
 import org.orecruncher.dsurround.lib.resources.ResourceUtilities;
+import org.orecruncher.dsurround.lib.seasons.ISeasonalInformation;
+import org.orecruncher.dsurround.lib.seasons.SeasonManager;
 import org.orecruncher.dsurround.lib.version.IVersionChecker;
 import org.orecruncher.dsurround.lib.version.VersionChecker;
 import org.orecruncher.dsurround.lib.version.VersionResult;
@@ -128,6 +129,8 @@ public final class Client {
                 .registerSingleton(ISoundLibrary.class, SoundLibrary.class)
                 .registerSingleton(IBiomeLibrary.class, BiomeLibrary.class)
                 .registerSingleton(IDimensionLibrary.class, DimensionLibrary.class)
+                .registerSingleton(IDimensionInformation.class, DimensionInformation.class)
+                .registerFactory(ISeasonalInformation.class, () -> SeasonManager.HANDLER)
                 .registerSingleton(IBlockLibrary.class, BlockLibrary.class)
                 .registerSingleton(IItemLibrary.class, ItemLibrary.class)
                 .registerSingleton(IEntityEffectLibrary.class, EntityEffectLibrary.class)
@@ -179,11 +182,6 @@ public final class Client {
         // Force instantiation of the core Handler. This should cause the rest
         // of the dependencies to be initialized.
         container.resolve(Handlers.class);
-
-        // Make sure our particle sheets get registered otherwise they will not render.
-        // These sheets are purely client side - they have to be manhandled into the
-        // Minecraft environment.
-        ParticleSheets.register();
 
         this.logger.info("[%s] Finalization complete", Constants.MOD_ID);
     }