+++ common/src/main/java/org/orecruncher/dsurround/config/libraries/IDimensionLibrary.java
@@ -1,7 +1,7 @@
 package org.orecruncher.dsurround.config.libraries;
 
 import net.minecraft.world.level.Level;
-import org.orecruncher.dsurround.config.dimension.DimensionInfo;
+import org.orecruncher.dsurround.config.DimensionInfo;
 
 public interface IDimensionLibrary extends ILibrary {
     DimensionInfo getData(final Level world);