+++ common/src/main/java/org/orecruncher/dsurround/lib/WeightTable.java
@@ -1,10 +1,10 @@
 package org.orecruncher.dsurround.lib;
 
-import org.orecruncher.dsurround.lib.collections.ObjectArray;
+import net.minecraft.util.random.WeightedEntry;
+import net.minecraft.util.random.WeightedRandom;
 import org.orecruncher.dsurround.lib.random.IRandomizer;
 import org.orecruncher.dsurround.lib.random.Randomizer;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Stream;
@@ -12,58 +12,33 @@ import java.util.stream.Stream;
 /**
  * Classic WeightTable for random weighted selection.
  */
-@SuppressWarnings("unused")
 public class WeightTable {
 
     public static <T> Optional<T> makeSelection(final Stream<? extends IItem<T>> inputStream) {
-        return makeSelection(inputStream.toList(), Randomizer.current());
+        return makeSelection(inputStream, Randomizer.current());
     }
 
     public static <T> Optional<T> makeSelection(final Stream<? extends IItem<T>> inputStream, IRandomizer randomizer) {
         return makeSelection(inputStream.toList(), randomizer);
     }
 
-    public static <T> Optional<T> makeSelection(final Collection<? extends IItem<T>> selections) {
-        return makeSelection(selections, Randomizer.current());
-    }
-
-    public static <T> Optional<T> makeSelection(final Collection<? extends IItem<T>> selections, IRandomizer randomizer) {
+    public static <T> Optional<T> makeSelection(final List<? extends IItem<T>> selections, IRandomizer randomizer) {
         if (selections.isEmpty())
             return Optional.empty();
 
-        if (selections.size() == 1) {
-            T theItem;
-            if (selections instanceof List<? extends IItem<T>> theList)
-                theItem = theList.getFirst().getItem();
-            else if (selections instanceof ObjectArray<? extends IItem<T>> theArray)
-                theItem = theArray.getFirst().getItem();
-            else
-                theItem = selections.iterator().next().getItem();
-            return Optional.of(theItem);
-        }
+        if (selections.size() == 1)
+            return Optional.of(selections.getFirst().data());
 
-        int totalWeight = 0;
-        for (var e : selections)
-            totalWeight += e.getWeight();
+        int totalWeight = WeightedRandom.getTotalWeight(selections);
+
         if (totalWeight == 0)
             return Optional.empty();
 
         int targetWeight = randomizer.nextInt(totalWeight);
-
-        for (var e : selections) {
-            targetWeight -= e.getWeight();
-            if (targetWeight < 0)
-                return Optional.of(e.getItem());
-        }
-
-        // Shouldn't get here
-        throw new RuntimeException("Bad weight table - ran off the end");
+        return WeightedRandom.getWeightedItem(selections, targetWeight).map(IItem::data);
     }
 
-    public interface IItem<T> {
-
-        int getWeight();
-
-        T getItem();
+    public interface IItem<T> extends WeightedEntry {
+        T data();
     }
 }