+++ common/src/main/java/org/orecruncher/dsurround/lib/seasons/compat/SereneSeasons.java
@@ -1,9 +1,10 @@
 package org.orecruncher.dsurround.lib.seasons.compat;
 
+import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.biome.Biome;
+import org.orecruncher.dsurround.config.libraries.IDimensionInformation;
 import sereneseasons.api.season.Season;
 import sereneseasons.api.season.SeasonHelper;
 import sereneseasons.season.SeasonHooks;
@@ -18,20 +19,23 @@ public class SereneSeasons extends AbstractSeasonProvider {
     private Season.TropicalSeason tropicalSeason;
     private Component computed;
 
-    public SereneSeasons() {
+    private final IDimensionInformation dimensionInformation;
+
+    public SereneSeasons(IDimensionInformation dimensionInformation) {
         super("Serene Seasons");
+        this.dimensionInformation = dimensionInformation;
     }
 
     @Override
-    public Optional<Component> getCurrentSeason(Level world) {
-        var helper = SeasonHelper.getSeasonState(world);
+    public Optional<Component> getCurrentSeason() {
+        var helper = SeasonHelper.getSeasonState(this.level());
         var subSeason = helper.getSubSeason();
         return Optional.of(Component.literal(subSeason.toString()));
     }
 
     @Override
-    public Optional<Component> getCurrentSeasonTranslated(Level world) {
-        var helper = SeasonHelper.getSeasonState(world);
+    public Optional<Component> getCurrentSeasonTranslated() {
+        var helper = SeasonHelper.getSeasonState(this.level());
         if (this.subSeason != helper.getSubSeason() || this.tropicalSeason != helper.getTropicalSeason()) {
             var subSeasonKey = "desc.sereneseasons." + helper.getSeason().toString().toLowerCase(Locale.ROOT);
             var tropicalSeasonKey = "desc.sereneseasons." + helper.getTropicalSeason().toString().toLowerCase(Locale.ROOT);
@@ -45,35 +49,42 @@ public class SereneSeasons extends AbstractSeasonProvider {
         return Optional.of(this.computed);
     }
 
-    public boolean isSpring(Level world) {
-        var helper = SeasonHelper.getSeasonState(world);
+    public boolean isSpring() {
+        var helper = SeasonHelper.getSeasonState(this.level());
         return helper.getSeason() == Season.SPRING;
     }
 
-    public  boolean isSummer(Level world) {
-        var helper = SeasonHelper.getSeasonState(world);
+    public  boolean isSummer() {
+        var helper = SeasonHelper.getSeasonState(this.level());
         return helper.getSeason() == Season.SUMMER;
     }
 
-    public  boolean isAutumn(Level world) {
-        var helper = SeasonHelper.getSeasonState(world);
+    public  boolean isAutumn() {
+        var helper = SeasonHelper.getSeasonState(this.level());
         return helper.getSeason() == Season.AUTUMN;
     }
 
-    public  boolean isWinter(Level world) {
-        var helper = SeasonHelper.getSeasonState(world);
+    public  boolean isWinter() {
+        var helper = SeasonHelper.getSeasonState(this.level());
         return helper.getSeason() == Season.WINTER;
     }
 
     @Override
-    public Biome.Precipitation getPrecipitationAt(Level world, BlockPos blockPos) {
-        var biome = world.getBiome(blockPos);
-        return SeasonHooks.getPrecipitationAtSeasonal(world, biome, blockPos);
+    public Biome.Precipitation getPrecipitationAt(BlockPos blockPos) {
+        var level = this.level();
+        var biome = level.getBiome(blockPos);
+        return SeasonHooks.getPrecipitationAtSeasonal(level, biome, blockPos);
     }
 
     @Override
-    public float getTemperature(Level world, BlockPos blockPos) {
-        var biome = world.getBiome(blockPos);
-        return SeasonHooks.getBiomeTemperature(world, biome, blockPos);
+    public float getTemperature(BlockPos blockPos) {
+        var level = this.level();
+        var biome = level.getBiome(blockPos);
+        return SeasonHooks.getBiomeTemperature(level, biome, blockPos);
+    }
+
+    @Override
+    public ClientLevel level() {
+        return this.dimensionInformation.level();
     }
 }