+++ common/src/main/java/org/orecruncher/dsurround/lib/seasons/compat/VanillaSeasons.java
@@ -2,7 +2,6 @@ package org.orecruncher.dsurround.lib.seasons.compat;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.level.Level;
 import org.orecruncher.dsurround.mixinutils.IBiomeExtended;
 
 import java.util.Optional;
@@ -14,18 +13,18 @@ public class VanillaSeasons extends AbstractSeasonProvider {
     }
 
     @Override
-    public Optional<Component> getCurrentSeason(Level world) {
+    public Optional<Component> getCurrentSeason() {
         return Optional.of(Component.translatable("dsurround.text.seasons.spring"));
     }
 
     @Override
-    public Optional<Component> getCurrentSeasonTranslated(Level world) {
-        return getCurrentSeason(world);
+    public Optional<Component> getCurrentSeasonTranslated() {
+        return getCurrentSeason();
     }
 
     @Override
-    public float getTemperature(Level world, BlockPos blockPos) {
-        var biome = world.getBiome(blockPos).value();
+    public float getTemperature(BlockPos blockPos) {
+        var biome = this.level().getBiome(blockPos).value();
         return ((IBiomeExtended)(Object)biome).dsurround_getTemperature(blockPos);
     }
 }